+++ tests/api/test_translations.py
@@ -1,8 +1,10 @@
 import graphene
 import pytest
+from django.contrib.auth.models import Permission
 
 from saleor.graphql.translations.schema import TranslatableKinds
-from tests.api.utils import get_graphql_content
+from saleor.graphql.translations.types import LanguageCodeEnum
+from tests.api.utils import assert_no_permission, get_graphql_content
 
 
 def test_product_translation(user_api_client, product):
@@ -1361,3 +1363,633 @@ def test_translations_query_inline_fragment(user_api_client, product):
 
     assert data["node"]["name"] == "Test product"
     assert data["node"]["translation"]["name"] == "Produkt testowy"
+
+
+QUERY_TRANSLATION_PRODUCT = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on ProductTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                product{
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "is_published, perm_codenames, return_product",
+    [
+        (True, ["manage_translations"], True),
+        (False, ["manage_translations"], False),
+        (False, ["manage_translations", "manage_products"], True),
+    ],
+)
+def test_translation_query_product(
+    staff_api_client,
+    product,
+    product_translation_fr,
+    is_published,
+    perm_codenames,
+    return_product,
+):
+    product.is_published = is_published
+    product.save()
+
+    product_id = graphene.Node.to_global_id("Product", product.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": product_id,
+        "kind": TranslatableKinds.PRODUCT.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_PRODUCT, variables, permissions=perms
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == product.name
+    assert data["translation"]["name"] == product_translation_fr.name
+    if return_product:
+        assert data["product"]["name"] == product.name
+    else:
+        assert not data["product"]
+
+
+QUERY_TRANSLATION_COLLECTION = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on CollectionTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                collection{
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "is_published, perm_codenames, return_collection",
+    [
+        (True, ["manage_translations"], True),
+        (False, ["manage_translations"], False),
+        (False, ["manage_translations", "manage_products"], True),
+    ],
+)
+def test_translation_query_collection(
+    staff_api_client,
+    collection,
+    collection_translation_fr,
+    is_published,
+    perm_codenames,
+    return_collection,
+):
+    collection.is_published = is_published
+    collection.save()
+
+    collection_id = graphene.Node.to_global_id("Collection", collection.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": collection_id,
+        "kind": TranslatableKinds.COLLECTION.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_COLLECTION, variables, permissions=perms
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == collection.name
+    assert data["translation"]["name"] == collection_translation_fr.name
+    if return_collection:
+        assert data["collection"]["name"] == collection.name
+    else:
+        assert not data["collection"]
+
+
+QUERY_TRANSLATION_CATEGORY = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on CategoryTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                category {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+def test_translation_query_category(
+    staff_api_client, category, category_translation_fr, permission_manage_translations
+):
+    category_id = graphene.Node.to_global_id("Category", category.id)
+
+    variables = {
+        "id": category_id,
+        "kind": TranslatableKinds.CATEGORY.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_CATEGORY,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == category.name
+    assert data["translation"]["name"] == category_translation_fr.name
+    assert data["category"]["name"] == category.name
+
+
+QUERY_TRANSLATION_ATTRIBUTE = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on AttributeTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                attribute {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+def test_translation_query_attribute(
+    staff_api_client, translated_attribute, permission_manage_translations
+):
+    attribute = translated_attribute.attribute
+    attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
+
+    variables = {
+        "id": attribute_id,
+        "kind": TranslatableKinds.ATTRIBUTE.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_ATTRIBUTE,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == attribute.name
+    assert data["translation"]["name"] == translated_attribute.name
+    assert data["attribute"]["name"] == attribute.name
+
+
+QUERY_TRANSLATION_ATTRIBUTE_VALUE = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on AttributeValueTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                attributeValue {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+def test_translation_query_attribute_value(
+    staff_api_client,
+    pink_attribute_value,
+    translated_attribute_value,
+    permission_manage_translations,
+):
+    attribute_value_id = graphene.Node.to_global_id(
+        "AttributeValue", pink_attribute_value.id
+    )
+
+    variables = {
+        "id": attribute_value_id,
+        "kind": TranslatableKinds.ATTRIBUTE_VALUE.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_ATTRIBUTE_VALUE,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == pink_attribute_value.name
+    assert data["translation"]["name"] == translated_attribute_value.name
+    assert data["attributeValue"]["name"] == pink_attribute_value.name
+
+
+QUERY_TRANSLATION_VARIANT = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on ProductVariantTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                productVariant {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "is_published, perm_codenames, return_variant",
+    [
+        (True, ["manage_translations"], True),
+        (False, ["manage_translations"], False),
+        (False, ["manage_translations", "manage_products"], True),
+    ],
+)
+def test_translation_query_variant(
+    staff_api_client,
+    product,
+    variant,
+    variant_translation_fr,
+    is_published,
+    perm_codenames,
+    return_variant,
+):
+    product.is_published = is_published
+    product.save()
+
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": variant_id,
+        "kind": TranslatableKinds.VARIANT.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_VARIANT, variables, permissions=perms
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == variant.name
+    assert data["translation"]["name"] == variant_translation_fr.name
+    if return_variant:
+        assert data["productVariant"]["name"] == variant.name
+    else:
+        assert not data["productVariant"]
+
+
+QUERY_TRANSLATION_PAGE = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on PageTranslatableContent{
+                id
+                title
+                translation(languageCode: $languageCode){
+                    title
+                }
+                page {
+                    id
+                    title
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "is_published, perm_codenames, return_page",
+    [
+        (True, ["manage_translations"], True),
+        (False, ["manage_translations"], False),
+        (False, ["manage_translations", "manage_pages"], True),
+    ],
+)
+def test_translation_query_page(
+    staff_api_client,
+    page,
+    page_translation_fr,
+    is_published,
+    perm_codenames,
+    return_page,
+):
+    page.is_published = is_published
+    page.save()
+
+    page_id = graphene.Node.to_global_id("Page", page.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": page_id,
+        "kind": TranslatableKinds.PAGE.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_PAGE, variables, permissions=perms
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["title"] == page.title
+    assert data["translation"]["title"] == page_translation_fr.title
+    if return_page:
+        assert data["page"]["title"] == page.title
+    else:
+        assert not data["page"]
+
+
+QUERY_TRANSLATION_SHIPPING_METHOD = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on ShippingMethodTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                shippingMethod {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "perm_codenames, return_shipping_method",
+    [
+        (["manage_translations"], False),
+        (["manage_translations", "manage_shipping"], True),
+    ],
+)
+def test_translation_query_shipping_method(
+    staff_api_client,
+    shipping_method,
+    shipping_method_translation_fr,
+    perm_codenames,
+    return_shipping_method,
+):
+    shipping_method_id = graphene.Node.to_global_id(
+        "ShippingMethod", shipping_method.id
+    )
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": shipping_method_id,
+        "kind": TranslatableKinds.SHIPPING_METHOD.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_SHIPPING_METHOD, variables, permissions=perms
+    )
+    content = get_graphql_content(response, ignore_errors=True)
+    data = content["data"]["translation"]
+    assert data["name"] == shipping_method.name
+    assert data["translation"]["name"] == shipping_method_translation_fr.name
+    if return_shipping_method:
+        assert data["shippingMethod"]["name"] == shipping_method.name
+    else:
+        assert not data["shippingMethod"]
+
+
+QUERY_TRANSLATION_SALE = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on SaleTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                sale {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "perm_codenames, return_sale",
+    [
+        (["manage_translations"], False),
+        (["manage_translations", "manage_discounts"], True),
+    ],
+)
+def test_translation_query_sale(
+    staff_api_client, sale, sale_translation_fr, perm_codenames, return_sale
+):
+    sale_id = graphene.Node.to_global_id("Sale", sale.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": sale_id,
+        "kind": TranslatableKinds.SALE.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_SALE, variables, permissions=perms
+    )
+    content = get_graphql_content(response, ignore_errors=True)
+    data = content["data"]["translation"]
+    assert data["name"] == sale.name
+    assert data["translation"]["name"] == sale_translation_fr.name
+    if return_sale:
+        assert data["sale"]["name"] == sale.name
+    else:
+        assert not data["sale"]
+
+
+QUERY_TRANSLATION_VOUCHER = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on VoucherTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                voucher {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+@pytest.mark.parametrize(
+    "perm_codenames, return_voucher",
+    [
+        (["manage_translations"], False),
+        (["manage_translations", "manage_discounts"], True),
+    ],
+)
+def test_translation_query_voucher(
+    staff_api_client, voucher, voucher_translation_fr, perm_codenames, return_voucher
+):
+    voucher_id = graphene.Node.to_global_id("Voucher", voucher.id)
+    perms = list(Permission.objects.filter(codename__in=perm_codenames))
+
+    variables = {
+        "id": voucher_id,
+        "kind": TranslatableKinds.VOUCHER.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_VOUCHER, variables, permissions=perms
+    )
+    content = get_graphql_content(response, ignore_errors=True)
+    data = content["data"]["translation"]
+    assert data["name"] == voucher.name
+    assert data["translation"]["name"] == voucher_translation_fr.name
+    if return_voucher:
+        assert data["voucher"]["name"] == voucher.name
+    else:
+        assert not data["voucher"]
+
+
+QUERY_TRANSLATION_MENU_ITEM = """
+    query translation(
+        $kind: TranslatableKinds!, $id: ID!, $languageCode: LanguageCodeEnum!
+    ){
+        translation(kind: $kind, id: $id){
+            __typename
+            ...on MenuItemTranslatableContent{
+                id
+                name
+                translation(languageCode: $languageCode){
+                    name
+                }
+                menuItem {
+                    id
+                    name
+                }
+            }
+        }
+    }
+"""
+
+
+def test_translation_query_menu_item(
+    staff_api_client,
+    menu_item,
+    menu_item_translation_fr,
+    permission_manage_translations,
+):
+    menu_item_id = graphene.Node.to_global_id("MenuItem", menu_item.id)
+
+    variables = {
+        "id": menu_item_id,
+        "kind": TranslatableKinds.MENU_ITEM.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_MENU_ITEM,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["translation"]
+    assert data["name"] == menu_item.name
+    assert data["translation"]["name"] == menu_item_translation_fr.name
+    assert data["menuItem"]["name"] == menu_item.name
+
+
+def test_translation_query_incorrect_kind(
+    staff_api_client, menu_item, permission_manage_translations
+):
+    menu_item_id = graphene.Node.to_global_id("MenuItem", menu_item.id)
+
+    variables = {
+        "id": menu_item_id,
+        "kind": TranslatableKinds.PRODUCT.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(
+        QUERY_TRANSLATION_MENU_ITEM,
+        variables,
+        permissions=[permission_manage_translations],
+    )
+    content = get_graphql_content(response)
+    assert not content["data"]["translation"]
+
+
+def test_translation_query_no_permission(staff_api_client, menu_item):
+    menu_item_id = graphene.Node.to_global_id("MenuItem", menu_item.id)
+
+    variables = {
+        "id": menu_item_id,
+        "kind": TranslatableKinds.MENU_ITEM.name,
+        "languageCode": LanguageCodeEnum.FR.name,
+    }
+    response = staff_api_client.post_graphql(QUERY_TRANSLATION_MENU_ITEM, variables)
+    assert_no_permission(response)