+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add sub-categories to random data - #4949 by @IKarbowiak
 - Deprecate `localized` field in Money type - #4952 by @IKarbowiak
 - Fix for shipping api doesn't apply taxes - #4913 by @kswiatek92
+- Query object translation with only manage_translation permission - #4914 by @fowczarek
 - Add cutomer note to draft orders api - #4973 by @IKarbowiak
+
 
 ## 2.9.0